NCBC-2993: Dispose of CTS within CancellationTokenPair
Motivation
----------
In cases where both an internal and external cancellation token is
supplied to the CancellationTokenPair and one of those tokens is
long-lived we will leak memory as more and more callbacks are
registered on the long-lived token. These callbacks take up memory
themselves and root CancellationTokenSource instances during GC.

Modifications
-------------
Rewrite CancellationTokenPair to be backed by a
CancellationTokenPairSource which implements IDisposable. Update all
call points to create a CancellationTokenPairSource and always dispose
of it.

Bake the concept of a timeout for the internal token directly into
CancellationTokenPairSource.

Improve ThrowIfCancellationRequested to always have the internal or
external token, not the global token, in the OperationCanceledException.

Add UnsafeRegister overloads for .NET Core 3.1 and later so we can use
them in the future for perf reasons.

Results
-------
So long as we properly dispose of the CancellationTokenPairSource any
memory leaks around callback registration are resolved.

Many call sites which were creating a CTS with a timeout to pass into
the internal token have been simplified.

We have added a small heap allocation for the
CancellationTokenPairSource, however this seems more than worth the cost
for the extra guarantees it provides. In particular, it abstracts the
disposal away from the CancellationTokenPair we're passing down the call
stack so that it's clear only the creator is responsible for disposal.
This is modeled on the pattern used by CancellationTokenSource and
CancellationToken.

CancellationTokenPair is now much smaller (8 bytes in 64-bit rather than
24 bytes) so there will be less copying between stack frames as it is
passed down the call stack. This is also modeled on the pattern used by
CancellationTokenSource and CancellationToken.

Change-Id: I9b2dd0bc91f107990c7ec47161dc84732ff1af9f
Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/165009
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>