NCBC-3007: Improve logging around connection pool scale down
Motivation
----------
There is some question that scale down of the DataFlowConnectionPool
may be causing operations to time out.

Modifications
-------------
Add logging in MultiplexingConnection around graceful shutdown waiting
for in-flight operations to complete.

Drop an unnecessary heap allocation and lock to control the dispose
of MultiplexingConnection, using Interlocked instead.

Re-enable logging of exceptions in the read loop that indicate
connection failure, but add a check to ensure that the exception is not
due to an intentional close operation and don't log those cases.

Add an integration test to cover connection pool scale down and ensure
that operations are not being dropped.

Extend ClusterFixture to allow specific tests to bootstrap using
specific options overriding the config.json file. Also, dispose of the
logger during tear down in case it needs to flush logs.

Change-Id: I76c708a925e2dc73a58f8e8156c59b1018fe4b97
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/165715
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard Ponton <richard.ponton@couchbase.com>