NCBC-3013: Keep connections alive after send is canceled
Motivation
----------
We are incorrectly killing active connections, dropping their in-flight
operations on the floor, if a requested operation send is canceled
while waiting on the OS to send.

Modifications
-------------
Ignore OperationCanceledExceptions on send.

Results
-------
Connections stay open if pending operation writes are cancelled, and
their in-flight operations will continue to process. If an operation
fails to send for other reasons and the connection is still closed.

Change-Id: I1989cd4094fc6348d0b31f09d95cb0fb9a38a8fa
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/165802
Reviewed-by: Richard Ponton <richard.ponton@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>