NCBC-3022: Improve efficiency of CID lookup
Motivation
----------
The lock being used when getting CID can cause timeouts if a large
number of operations are sent at once before the CID is populated.

Modifications
-------------
Rather than every operation independently waiting on an asynchronous
lock which is held for the duration of the GET_CID operation, hold a
very brief synchronous lock to create a single task.

Also fix some issues with integration tests.

Results
-------
When multiple operations want to refresh the CID simultaneously, they
will share and wait on the same task. When the task completes, all of
the operations are released to continue processing immediately. The
previous behavior would have each of them gain the lock via an async
continuation and then release it for the next operation.

This also significantly reduces heap allocations surrounding
simultaneous requests to get the CID.

Change-Id: I9a51c82164c1f534f5369a242aa8850b67c927df
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/166104
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Richard Ponton <richard.ponton@couchbase.com>