NCBC-2991: Support named bucket/collection DI in .NET 6
Motivation
----------
Starting in .NET 6 the methods we're using to apply strong name
signatures to the dynamic assembly throw PlatformNotSupportedException.

Modifications
-------------
Starting with .NET 5 don't strong name the dynamic assembly and drop the
InternalsVisibleTo attributes.

Results
-------
This is tested to work with a signed DI assembly in both .NET 5 and 6.
They both don't care that the unsigned dynamic assembly is referencing
a signed DI assembly. They also both accept the
IgnoresAccessChecksToAttribute attribute we're adding to the dynamic
assembly to access internal members of the DI assembly.

Behaviors are unchanged for older frameworks.

Change-Id: I469bde5dc4ee4f30dd58fbeab773a1025332211f
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/166191
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard Ponton <richard.ponton@couchbase.com>