NCBC-3021: Get memcached buckets working again
Motivation
----------
There is currently a regression where legacy memcached buckets are
completely non-functional.

Modifications
-------------
Rework the default indexer in BucketBase to be non-virtual and always
call Scope(string scopeName) so we can have a consistent method to
override once for all code paths and bucket types.

Rewrite the override in MemcachedBucket to call the base implementation
for the default scope.

Cache the action for creating the scope passed to the
ConcurrentDictionary for a minor performance bump.

Remove the unused LoadDefaultScope code path in BucketBase.

Added an extra integration test for memcached buckets.

Results
-------
Code is better organized, and the default scope is actually created for
memcached buckets the first time it is accessed. We should also see a
very small perf improvement getting scopes from eliminating a heap
allocation.

Change-Id: Ia58a6b830acc25f3099bcaef780b23ca8c70e92d
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/166237
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>