NCBC-2600: Default query HTTP Idle timeout to 4.5s
Motivation
----------
Changes to the server in 6.0.4 and later as a mitigation for slowloris
attacks mean that HTTP connections maybe closed if idle for more than
5s. This patch sets the client HTTP idle timeout to 4.5s and sets the
default ClusterOptions TCP Keep Alive values on the connection as well.

Modification
------------
 - Set the TCP Idle connect to 4.5s on HTTP services
 - Make the TCP pooling lifetime property for HTTP services
 - Set TCP Keep Alives on HTTP connections if enabled.

Result
------
The client should not receive IOException: the reponse ended prematurely
using the default values.

Change-Id: Ifb8060e331345f46c5058776cdd327b3ac9cad29
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/166276
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard Ponton <richard.ponton@couchbase.com>