NCBC-3019: Allow SSL cipher configuration - Couchbase database (NOTE: Created based on Review 166541 - Comments addressed + Merge conflicts resolved)
Motivation
----------
Add options to control the SSL Protocol and TLS Cipher Suite (TLS Cipher Suite is .Net Core App 3.1+ only) for SSlStream and HttpClient connections.  If the option(s) are not explicitly set, then defaults / no specific input parameters are used in the connection calls.

The SSL Protocol change is needed to support security configurations in environments that may be more strict (example: only TLS 1.2/1.3 allowed).

The TLS Cipher Suite change is needed for full/flexible .NET 5+ support because of security related changes made to the .NET framework: https://docs.microsoft.com/en-us/dotnet/core/compatibility/cryptography/5.0/default-cipher-suites-for-tls-on-linux

Modifications
-------------
Added Cluster Options:
  - EnabledSslProtocols (Type: SslProtocols)
  - EnabledTlsCipherSuites (Type: List<TlsCipherSuite>) [Available to .Net Core App 3.1+ Only]

If the above options are set, they are used in ConnectionFactory.cs and CouchbaseHttpClientFactory.cs in the SSLStream and HttpClient calls.

If EnabledSslProtocols is not set, a default value of "SslProtocols.Tls | SslProtocols.Tls11 | SslProtocols.Tls12" is used (consistent with the previous CB SDK configuration).

If EnabledTlsCipherSuites is not set (empty or NULL), the SSLStream and HttpClient calls do not pass in any specific Cipher Suites, so the defaults from the .NET framework is used.

Result
------
Without the change, using the CB SDK in certain environments may encounter a Bootstrap error if the protocols/security configuration requested does not match the allowed/supported ones from the environment.  With the change, CB SDK users can configure custom settings to match their environment/security needs, and the Bootstrap errors will be avoided.


Change-Id: I8f49a4578127442d4bea0f646eb8f021c25d2948
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/166625
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Brant Burnett <bburnett@centeredgesoftware.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>