NCBC-3032: Support Memory<byte> with RawBinaryTranscoder
Motivation
----------
Supporting encoding/decoding binary data as Memory<byte> may help in
some high-performance scenarios. For encoding, it allows the use of a
subset of a buffer without allocating/copying a new buffer. For
decoding, it allows the use of buffers from the ArrayPool.

Modifications
-------------
Accept both `Memory<byte>` and `ReadOnlyMemory<byte>` for encoding with
the RawBinaryTranscoder.

Allow `IMemoryOwner<byte>` to be requested when decoding via
`.ContentAs<IMemoryOwner<byte>>`, which returns a block of memory which
the caller is now responseible for disposing.

Results
-------
Consumers may opt-in to these high-performance paradigms as needed when
handling large binary objects.

Note: If the consumer fails to dispose the IMemoryOwner<byte>, there is
no risk of a memory leak. However, it may reduce GC performance due to
creating large holes in the heap requiring compaction.

Change-Id: Id66dd87b759ce6acd93bdc1eefa02b69021aaa33
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/166845
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>