NCBC-3033: Remove finalizer from OperationBase
Motivation
----------
Objects with finalizers reduce GC performance by pushing objects to
Gen 1 rather than collecting them in Gen 0. OperationBase can be managed
without a finalizer.

Modifications
-------------
Remove the finalizer from OperationBase, and go with a simple Dispose
mechanism instead.

Ensure all call sites which create an operation are disposing of the
operation correctly.

Results
-------
For call sites which were not disposing of the operation, the operation
will no longer live on into Gen 1 garbage collection. In cases where
ExtractData() was not called (i.e. error scenarios) the buffer will be
returned to the ArrayPool for reuse much sooner.

For all call sites CPU cycles related to managing the finalizers and the
finalizer queue have been eliminated.

For any call sites which were missed and are still not disposing (there
shouldn't be any, but you never know) there may still be a GC impact.
This is because the buffers in error scenarios will be promoted to at
least Gen 1 and then collected, leaving a hole in the heap eventually
requiring compaction.

Change-Id: I5e7152a644991126dd73425e6d0d14a6f980147e
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/166873
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>