NCBC-3034: Reduce blocking in async methods in data structures
Motivation
----------
Blocking within async methods to wait on other async results is an
anti-pattern with significant performance impact.

Modifications
-------------
Add CreateBackingStoreAsync to PersistentStoreBase and replace all call
sites with the async overload. Mark the old method as obsolete.

Make the inherited types sealed so they may take advantage of
devirtualization. This is safe despite the fact they are public because
the constructors are internal so no consumer could inherit from them.

Optimize ClearAsync by avoiding an unnecessary Insert before the Upsert
and then marking BackingStoreChecked as true.

Make CountAsync actually asynchronous on all types, and .Keys and
.Values on PersistentDictionary.

Fix a bug in PersistentList where setting an index was actually
inserting instead of setting.

Fix a bug in PersistentDictionary.AddAsync where it uses the document
ID rather than the key for adding.

Flesh out integration tests.

Results
-------
More efficient thread utilization when calling the Async overloads on
DataStructures.

A couple of bugs have been fixed that were discovered in the process of
fleshing out the tests.

Change-Id: Ia17b45f71f669c61b10f2ccb35dfe992657ccb6a
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/166946
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>