NCBC-3035: Improve performance and trimmability of EnumExtensions
Motivation
----------
EnumExtensions helpers are not very performant, repeating reflection on
every call and unnecessarily boxing.

Modifications
-------------
Make use of generics for GetDescription to cut out unnecessary boxing.

Add a static caching layer for the results to avoid repeating
reflection on every call.

Enable nullable reference types for the extensions and add some
null-forgiving operators where required.

Annotate the methods with trimming attributes, and add the trimming
attributes as internal attributes for older frameworks. These are
placed in a shared location so they may be used by any of the projects.

Results
-------
Improved performance of EnumExtensions method calls.

Using generics now allows us to properly annotate for trimming. The
attributes will have no effect on pre-5.0 frameworks, but we include
them as internals to avoid excessive conditional compilation directives.
This still doesn't enable trimming support, but is a step in that
direction.

Change-Id: I8926e3e03a336aa36d5b8b5357946b76edf2ee40
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/166990
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>