NCBC-3036: Add tracing spans for compression/decompression
Motivation
----------
When implementing compression, it's important to understand how
compression is affecting your performance profile. A child tracing span
for the compression and decompression phases will provide this with
observability.

Modifications
-------------
Pass parent spans to IOperationCompressor and create child spans for
both compression and decompression.

Ensure tracing strings are readonly or const.

Rewrite logging in OperationCompressor to use LoggerMessage.Define.

Add an integration test to Otel that shows compression/decompression
spans.

Results
-------
Compression and decompression durations are now observable in traces.

Reduced boxing and heap allocations when logging for compression.

Change-Id: Icd07a4827d68534abd53a63577cc6e11de2335de
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/167047
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>