NCBC-3044: Provide a way to wire to an external TracerProviderBuilder
Motivation
----------
The current pattern for wiring OpenTelemetry tracing to the SDK assumes
that the SDK is creating a TracerProvider. Typically, this is actually
done externally, and there is one TracerProvider built for the
application. We should instead provide a means to wire the SDK to an
unbuilt TracerProviderBuilder.

Modifications
-------------
Mark the existing constructor of OpenTelemetryRequestTracer as obsolete
and add a new overload that doesn't build a TracerProvider.

Add an extension method to subscribe a TracerProviderBuilder to the
Couchbase event source.

Check the parent span for a null Id before constructing new activities,
as the parent span may have a null Activity if sampling is enabled.

Correct the span Parent property to actually return the parent.

Add some unit tests to ensure parent span behaviors.

Results
-------
A consumer may now register the SDK with an OpenTelemetry provider which
is being managed outside the SDK. Some potential null-ref bugs have also
been resolved.

Change-Id: I6ac8f4e272c0e4c4762837b79dcb3cfc71b9f512
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/167288
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard Ponton <richard.ponton@couchbase.com>