NCBC-3047: Stop tracing on cluster dispose
Motivation
----------
When tracing is started (by ClusterOptions.BuildServiceProvider) it is
never stopped. This could cause issues in production, but is, in
particular, interfering with unit tests for tracing.

Modifications
-------------
Move tracing start logic to ClusterContext.Start rather than
BuildServiceProvider.

Dispose of the trace listeners when the ClusterContext is disposed if
they were internally manufactured.

Results
-------
Reduced side-effects and improved separation of concerns by moving
the trace start to ClusterContext.Start.

For internally manufactured listeners they will be cleaned up properly
when the cluster is disposed. This will prevent a memory leak in the
(clearly non-standard) scenario where the cluster is disposed and
reopened repeatedly.

For externally created listeners the consumer is still expected to
manage the lifetime of the listener.

Change-Id: I15982a3ff5cf0aefed23e1a909f0211266755e82
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/167450
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>