NCBC-3049: Prevent logs after tests complete
Motivation
----------
We're having sporadic failurs of unit tests due to logging to the test
output helper after the test completes. This seems to be caused by tasks
remaining active after the test completes. For example,
GetAnyReplicaAsync creates multiple requests that run in parallel, but
returns when the first one completes. Therefore, the test may complete
while the other requests are still active and potentially logging.

Modifications
-------------
Stop writing to ITestOutputHelper once the XUnitLoggerListener is
disposed.

Ensure that consumers of XUnitLoggerListener dispose of their
ILoggerFactory after each test.

Results
-------
Additional logs after the test completes should be dropped and not
attempt to write to the ITestOutputHelper.

Change-Id: Ic9f3c89f50cd4054790098f360f7af18a75f011d
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/167502
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>