NCBC-2829: NOOP operations fail with ObjectDisposedException
Motivation
----------
Handles an ODE in the ping reports if a node is unreachable. The actual
error will be logged, but this is a side effect in that ping report is
on a timer and will try to keep using a disposed socket. We already know
the reason for the initial exception and logging this only creates
confusion as it is not the root cause.

Change-Id: Id23f2f64d5cde614e176d118fd6e982237a6a535
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/167881
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard Ponton <richard.ponton@couchbase.com>
Reviewed-by: Brant Burnett <bburnett@centeredgesoftware.com>