NCBC-3059: Eliminate OperationCancellationRegistration heap allocs
Motivation
----------
A significant percentage of our heap allocations are coming from the
OperationCancellationRegistration objects allocated for every operation
(29.8MB for 300k GET operations, for example).

Modifications
-------------
Store the CancellationTokenPair for each attempt rather than just the
CancellationToken on the IOperation. Do this when it's registered rather
than as part of the RetryOrchestrator.

Convert the OperationCancellationRegistration to a struct, and avoid
boxing by passing the IOperation as state to the cancellation callbacks.

Results
-------
The OperationCancellationRegistration is never created on the heap or
boxed. In exchange, OperationBase is 8 bytes larger and the stack frame
for the method call using OperationCancellationRegistration (or the
boxed async state machine) is 24 bytes larger (both examlpes based on a
64-bit process).

Given that we are dropping the allocation of an object on the heap that
was 48 bytes + a header, this tradeoff is worth it.

Change-Id: I6e8a4b44a7d1691f9d5d7a89875a15aee9a65a76
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/167901
Reviewed-by: Richard Ponton <richard.ponton@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>