NCBC-3060: Reduce heap allocations for ClusterNode.SendAsync
Motivation
----------
Reduce garbage collection required when performing key/value operations.

Modifications
-------------
For cases where a circuit breaker is in use, switch to use a simple
async/await instead of the overly complicated ContinueWith that was
performing extra closures and allocating an ExecutionContext in order
to prevent capture.

Move half-open circuit breaker tests to separate methods that use the
UnsafeQueueUserWorkItem as a cheaper alternative to suppressing the
ExcecutionContext and calling Task.Run.

Results
-------
The code is cleaner and easier to understand and performs fewer heap
allocations.

Marking success/failure on the circuit breaker now occurs in the
flow before the caller is informed of the error. However, this is a
very inexpensive operation. Half-open tests are still queued
separately.

Change-Id: Id34fca0dbf844fbc192d6ece33e1d0498bace0df
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/167925
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>