NCBC-3062: Don't set ServicePoint.MaxIdleTime for netcoreapp3.1
Motivation
----------
Setting this on the ServicePoint is redundant in netcoreapp3.1, it's
being set on the SocketsHttpHandler.

Modifications
-------------
Don't set the MaxIdleTime when targeting netcoreapp3.1 (it was already
excluded from net5.0 and later).

Drop the ServicePoint.ConnectionLimit setting altogether, as this is
already set for all target frameworks on either the HttpClientHandler or
SocketsHttpHandler.

Results
-------
Redundant settings, and the possible ArgumentOutOfRange exceptions which
may result, are no longer included in netcoreapp3.1 targets.

Change-Id: If3fc11ba3dcb907528c0803ab39fda9a8d256406
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/168000
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>