NCBC-3063: Replace Stopwatch in AsyncState with a lightweight approach
Motivation
----------
Currently, every K/V operation is allocating two Stopwatch objects on
the heap. One is used to track the run time of the operation and needs
sub-millisecond resolution. However, the other is used purely to
identify AsyncState objects which are orphaned and for which a response
has never been received, and high resolution is not required.

Modifications
-------------
Create the LightweightStopwatch structure as an abstraction over two
different stopwatch implementations. For older .NET frameworks, continue
to use the Stopwatch class. For newer frameworks, use
Environment.TickCount64.

Switch MultiplexingConnection and AsyncState to use the new
LightweightStopwatch structure.

Results
-------
For .NET Core 3.1 and later, AsyncState and MultiplexingConnection no
longer allocate a Stopwatch on the heap. The cost of this is a lower
resolution on the timer (10-16 milliseconds, per MS docs). However,
these timers are only used for detecting idleness on the scale of
minutes, high resolution is not required.

There is no benefit or behavioral change for older frameworks.

Note: This is modeled on the .NET approach used for tracking idle times
within SocketsHttpHandler.

Benchmark comparing creating and checking elapsed time on a default
Stopwatch vs the new LightweightStopwatch:

BenchmarkDotNet=v0.13.1, OS=Windows 10.0.22000
Intel Core i7-9750H CPU 2.60GHz, 1 CPU, 12 logical and 6 physical cores
.NET SDK=6.0.101
  [Host]     : .NET 6.0.1 (6.0.121.56705), X64 RyuJIT
  Job-CLZYKM : .NET 6.0.1 (6.0.121.56705), X64 RyuJIT

Runtime=.NET 6.0  Toolchain=net6.0

|      Method |      Mean |     Error |    StdDev |    Median | Ratio | Rank |  Gen 0 | Allocated |
|------------ |----------:|----------:|----------:|----------:|------:|-----:|-------:|----------:|
|     Default | 31.304 ns | 0.5365 ns | 0.9255 ns | 30.809 ns |  1.00 |    2 | 0.0063 |      40 B |
| Lightweight |  6.187 ns | 0.1510 ns | 0.2395 ns |  6.166 ns |  0.20 |    1 |      - |         - |

Change-Id: I5d9e81a9355d63ae42b1ab4ecf77c560720c807d
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/168020
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard Ponton <richard.ponton@couchbase.com>