NCBC-3081: Reuse CancellationTokenSource in .NET 6
Motivation
----------
.NET 6 offers a system to allow reuse of a CancellationTokenSource so
long as it hasn't been canceled. We should use this to reduce heap
allocations where we allocate a CTS to manage a timeout.

Modifications
-------------
Create a CancellationTokenSourcePool class which implements a low-cost
pool of reusable CancellationTokenSource objects. The class is a no-op
for older frameworks which don't support TryReset.

Use the pool for query timeouts in RetryOrchestrator and for K/V
timeouts in CancellationTokenPairSource.

Results
-------
For .NET 6, heap allocations of CancellationTokenSource should be
significantly reduced. Up to Environment.ProcessorCount*2 instances will
be persisted for reuse (based on current defaults in .NET object pool).
Note that this is just the maximum number of idle CTS objects in the
pool, under load there will likely be many more moving in and out of the
pool being reused.

If a timeout does occur, TryReset will return false and the CTS will be
dropped from the pool.

With inlining, this should have no impact on older frameworks.

Change-Id: I67133548ff8f678508a844e9cd2ff1c7224ac90c
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/168407
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard Ponton <richard.ponton@couchbase.com>