[WIP] NCBC-3092: Resolve DNS for each connection rather than node bootstrap
Motivation
----------
If the IP address of a node changes during a failover event but the
DNS name stays the same the SDK is unable to reconnect to the node. In
particular, this may happen with the Couchbase Autonomous Operator
in the case of a Kubernetes node failure, node drain, or during a
rolling Kubernetes upgrade.

Modifications
-------------
Refactor ClusterNode and ClusterNodeCollection to be keyed by a
HostEndpointWithPort instead of an IPEndPoint. HostEndpointWithPort is
like the existing HostEndpoint structure but the port is required rather
than optional. Drop the BootstrapEndpoint property since it's
unnecessary now.

Change many types to pass down HostEndpointWithPort instead of
IPEndPoint. This flows all the way to the connection, which finally
resolves the IP using IIpEndPointService when the connection is being
established. Key mapping also returns a HostEndpointWithPort rather than
an IPEndPoint.

Remove node-related overloads and caching from IpEndPointService.

Make several factory types synchronous now that they don't need to
resolve IP addresses so early in the process.

Add a version of GetRandom named GetRandomValueType that will return
null for value types if the list is empty. The previous implementation
would return the default value.

Results
-------
DNS names are resolved as each connection is created rather than only
the first time the node is seen in configuration. DNS caching is
provided by the OS, meaning TTL will be respected.

Several code paths around VBucketServerMap and key mapper factories are
also much cleaner and synchronous now that the factories don't need to
perform DNS resolution so early in bootstrap.

Change-Id: Icea7a9502542aa39bd0f6553994e42a7be83d6cb
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/169149
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>