NCBC-3095: Shutdown the Bootstrapper loop on Dispose
Motivation
----------
The bootstrapper loop continues to run indefinitely after the cluster or
bucket is disposed.

Modifications
-------------
Refactor to use some better practices around long-running task
management (only suppress flow if not already suppressed, instead of
Task.Run start an async method and discard the returned task).

Cancel the CancellationTokenSource before disposing it, Dispose does not
set the state to cancelled or fire registered cancellation callbacks.

Use an Interlocked.Exchanged to be extra cautious that the Dispose
logic is only executed once.

Results
-------
The bootstrapper loop is properly stopped when the cluster or bucket is
disposed.

Change-Id: I2d03bea9fd5f47dcb34cf1b015282f30e717d263
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/169289
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>