NCBC-3096: Cleanup after memcached buckets are removed
Motivation
----------
Memcached buckets may leak a CancellationTokenRegistration and
HttpStreamingConfigListener instance if the bucket is removed from the
Couchbase cluster. They would persist until the entire Cluster is
disposed.

Modifications
-------------
Don't register the dispose of HttpStreamingConfigListener on the
CancellationTokenSource. Instead, dispose of all of them directly in the
Dispose method.

Cleanup the pattern used for the long-running background task.

Results
-------
The CancellationTokenRegistration is no longer held after an Unsubscribe
on a memcached bucket. This in turn releases all references to the
HttpStreamingConfigListener so it may be collected as well.

Change-Id: I39d830c5717dd864b1d0032d644c9b66f14bedf2
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/169291
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>