NCBC-3109: Quota Limited Exceptions not being thrown for some Management apis
Motivations:
When rate limited errors are returned from the server when doing some SearchIndexes operation,
the SDK does not parse them as such, instead returning a geneirc CouchbaseException. And
also similarly for some collection management operations, like exceeding the max number of
collections on a scope, a generic Couchbase Exception is thrown.

Modifications:
- Additional HttpResponseError message parsing to convert the errors into
  QuotaLimitedExceptions
- Added some more rate limit tests for Search, SearchIndexManager and CollectionManager to
  cover these cases

Change-Id: I262c7b0062818d2cf255b3275512ffab00750ae2
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/169760
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>