NCBC-3126: Optimize vBucket map deserialization
Motivation
----------
There are currently a lot of heap allocations when deserializing
vBucket maps due to list resizing. Given that we have advance knowledge
about the structure of the maps we can optimize this.

Modifications
-------------
Add a custom JSON serializer for vBucket maps that is optimized for
1024 vBuckets and assumes each vBucket has the same number of replicas.
This deserializer will still function with other lengths, but will
fallback to less efficient paradigms in those cases.

Also fixed a compiler error in JSON stream reading found with the
slightly stricter compiler in VS 2022.1.

Results
-------
Drastically reduced heap allocations deserializing BucketConfig objects.

Change-Id: Ib64e82c33bfa0e00182b65049782ea6930002065
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/171468
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>