NCBC-3133: Switch obsolete OperationResult types to use System.Text.Json
Motivation
----------
We're switching all internals to use System.Text.Json.

Modifications
-------------
Mark IOperationResult and OperationResult types obsolete as they are not
used at all by the SDK but are public.

Since we're touching them, go ahead and turn on nullable ref types.

Switch ToString to use System.Text.Json. However, drop the inclusion of
Content since it can't be included in a way safe for link-level
trimming.

Results
-------
OperationResult is now clearly marked as Obsolete so it may be removed
in the future. System.Text.Json is now used for serialization in calls
to ToString().

Change-Id: I22f81d438d24c85723445f26fd1e0376001942a8
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/171826
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>