NCBC-3177: NRE when rebalancing and cluster map is missing an alternate address
Motivation
----------
If a node with alternate addresses is swapped out and then added back
into the cluster, it will be reset back to its 'fresh' state and the
alternate address setting will be removed. If not added back in the SDK
will expect the alternate address but will instead throw a NullRef
exception. This commit handles the null case so that the the NRE will
not be thrown. The alternate addresses will set still need to be enabled
on the server for the SDK to behave correctly.

Modifications
-------------
 - Handle the null case if a node does not contain Alternate Addresses
   in BucketConfig.

Change-Id: I5e9ef88e08d51fda43d496c40cece6fd9725bf02
Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/173295
Reviewed-by: Richard Ponton <richard.ponton@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>